Adds modyfies utils.scala in 2.4, updates all jars, it finally starts a Spark 4 session
