+++ R/spark_compile.R
@@ -1,6 +1,6 @@
 sparklyr_jar_spec_list <- function() {
   list(
-    list(spark = "2.4.0", scala = "2.11"),
+    #list(spark = "2.4.0", scala = "2.11"),
     list(spark = "2.4.0", scala = "2.12"),
     list(spark = "3.0.0", scala = "2.12"),
     list(spark = "3.5.0", scala = "2.12", jar_name = "sparklyr-master-2.12.jar"),
